+++ CHANGELOG.md
@@ -1,7 +1,26 @@
 # Changelog
 
+## [2.0.0] - 2023-07-17
+### Added
+ - new fault_name model & queries
+### Changed
+ - solvis dependency >= 0.7.0
+ - API changes
+    - get_ruptures -> query.get_location_radius_ruptures
+    - get_rupture_ids -> query.get_location_radius_rupture_ids
+
+## [1.1.0] - 2023-04-24
+### Added
+ - poetry run cli -> scripts/cli.py
+ - for extra dependencies install with `poetry install --with scripts`
+ - get_ruptures query
+ - get_rupture_ids query
+
+### Changed
+ - requires nzshm-model >=0.3.0
+
 ## [1.0.1] - 2023-04-19
-## Removed
+### Removed
  - geopandas dependency
 
 ## [1.0.0] - 2023-04-19